Use checkstyle for code style variations
