--- src/main/java/Ponder_Pika/Exception/PonderPikaException.java
@@ -1,13 +1,0 @@
-package Ponder_Pika.Exception;
-
-public class PonderPikaException extends Exception{
-
-    public PonderPikaException(String message){
-        super(message);
-    }
-
-    @Override
-    public String toString() {
-        return "Oh pika pi (T_T) - " + this.getMessage();
-    }
-}
