--- src/main/java/Ponder_Pika/Main.java
@@ -1,8 +1,0 @@
-package Ponder_Pika;
-
-public class Main {
-    public static void main(String[] args) {
-        Ponder_Pika bot = new Ponder_Pika();
-        bot.echo();
-    }
-}
