+++ src/main/java/ponderpika/Main.java
@@ -1,0 +1,15 @@
+package ponderpika;
+
+/**
+ * The entry point for the PonderPika application.
+ * <p>
+ * The Main class contains the main method, which initializes the PonderPika
+ * application and starts its execution.
+ * </p>
+ */
+public class Main {
+    public static void main(String[] args) {
+        PonderPika bot = new PonderPika();
+        bot.echo();
+    }
+}
