+++ src/main/java/ponderpika/task/Deadline.java
@@ -1,14 +1,14 @@
-package Ponder_Pika.Task;
+package ponderpika.task;
 
-import java.time.format.DateTimeFormatter;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
 
 /**
  * This class represents a specific type of task that has a deadline.
- * It extends the {@code Task} class and includes additional information about the deadline
+ * It extends the {@code task} class and includes additional information about the deadline
  * by which the task must be completed.
- * The {@code Deadline} class overrides the following methods from the {@code Task} class:
+ * The {@code Deadline} class overrides the following methods from the {@code task} class:
  * saveFullDetails() and toString()
  */
 public class Deadline extends Task {
