+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ----------------------------------------------------------
-Hello I'm Ponder_Pika
+Hello I'm ponderpika
 
 It is a great day to ponder! How may I help you?
 ----------------------------------------------------------
